+++ Sources/DeclarativeUIKit/ViewExtension/UIActivityIndicatorView+.swift
@@ -1,30 +1,35 @@
-//
-//  UIActivityIndicatorView+.swift
-//  
-//
-//  Created by sakiyamaK on 2022/04/14.
-//
-
 import UIKit
 
-//MARK: - Declarative method
 public extension UIActivityIndicatorView {
+
+    @discardableResult
+    func style(_ style: UIActivityIndicatorView.Style) -> Self {
+        self.style = style
+        return self
+    }
 
     @discardableResult
     func hidesWhenStopped(_ hidesWhenStopped: Bool) -> Self {
         self.hidesWhenStopped = hidesWhenStopped
         return self
     }
-    
+
     @discardableResult
-    func color(_ color: UIColor) -> Self {
+    func color(_ color: UIColor?) -> Self {
         self.color = color
         return self
     }
-    
+}
+
+public extension UIActivityIndicatorView {
+
     @discardableResult
-    func style(_ style: Style) -> Self {
-        self.style = style
+    func animating(_ isAnimating: Bool) -> Self {
+        if isAnimating {
+            self.startAnimating()
+        } else {
+            self.stopAnimating()
+        }
         return self
     }
 }
